+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/networks/Eth2Network.java
@@ -21,6 +21,7 @@ public enum Eth2Network {
   MAINNET,
   PRATER,
   KINTSUGI,
+  KILN,
   // Test networks
   MINIMAL,
   SWIFT,